Contact entries will now show up on one line regardless of length
Contacts with long names will now show up on one line as well.

There is a bug left (will be filed) where the truncation behaviour of
the contact names is a bit odd (the screen-width portion of the name is
all that is visible, however, the actual name div extends past the
visible region and therefore we cannot use normal methods to handle that
such as adding an ellipsis at the end).