Restructure layout to better respond to responsiveness
This makes a few changes to make sure that using responsive positioning
will work better.

The top html element is given min-height of 100% to take care of the
window being larger than the content.

The body is given a bottom margin to hold any footer content (60px is
used as the size for this, that would put the footer at the bottom with
very little space above it).

The pseudo-table layout for the buddy list is completely scrapped, it
was too hard to work with.  To make this work, the status display at the
bottom of the list is converted to an absolutely positioned footer.