Restructure contact list DOM
The actual goal here is to fix the .accessIcon images on the contact
list (they were not getting properly aligned to the right side) and to
fix the width of the instances div so that the bar would stretch across
the full screen.  This takes apart the structure that was there before
and re-implements it in a way that flows nicely.  Truncation of too-long
usernames is now also handled.

This also changes ids to classes because that feels a bit more sane.