Add a handler to send URL data to the copy+paste connection
This adds a handler to chrome that will block loads of
uproxy.org/(request|offer)/{payload} and instead sends the payload data
to the copypaste instance in the core.

The actual handling of the tab interception is done in browser-specific
areas while the parsing is handled in the generic_ui component.

For Chrome, when the request is intercepted, the tab is redirected to
show the standard app interface.  This behaviour can be a bit odd (e.g.
if the app is already open), however, it seems equally as good as the
alternative of showing a landing page just indicating that the url is
loaded.

For Firefox, no action is taken on the request and the UProxy window is
opened.