Set up the core components required for copy+paste
This creates a new class, RemoteConnection, that handles a direct
connection between two peers (getting and sharer) with signaling mesages
exchanged over some arbitrary interface.  Obviously, the main intention
of this is for the copy+paste component of the app where users can send
messages solely through copy+pasting the signalling messages, however,
it should be able to be used as a platform for other components as well.

This brings up what is probably my strangest decision in the design of
the component, the sendUpdate argument of the constructor.  This is
mapped directly to ui.update for the global copypaste instance, however,
the intention is that you would eventually be able to use that to build
the current infrastructure (remote-instance) on top of this class (I
think the ideal solution would be the ability to register callbacks
using something like connection.on(SIGNALLING_MESSAGE), however, that
does not seem to be possible).

There is currently no attempt whatsoever to provide deduplication
between this and remote-instance, that will most likely be taken care of
in future commits.