Update RemoteInstance to be backed by RemoteConnection
This change updates the RemoteInstance class to use RemoteConnection to
back the instance instead of directly accessing socksToRtc and rtcToNet.
There was an incredible level of code duplication between the classes,
so this makes a lot more sense in the long run.

When a RemoteInstance is created, we set up a new RemoteConnection in
that will handle all the backend functionality (the RemoteInstance still
handles things such as checking for permissions since that's not needed
for all connections).  We register an update function to it that checks
the type of each update and does an appropriate action for that.  For
most of them, this just means proxying the callback to ui.update,
switching the type, and adding an instanceId to it.  For state, we cache
the values locally.  Other than that, it's a pretty straightforward
conversion since most of the code in RemoteConnection is essentially a
copy-paste of the RemoteInstance code.

Testing changes:

The fakeSocksToRtc object we use now actually records when handlers are
bound to it and will fire the stopped handler when stop is called.  The
RemoteConnection code relies on the callback actually being triggered so
that was causing test failures.

For the generic_ui tests, we now spy on UserInterface::showNotification
since the code to trigger one of the notifications was moved out of the
core module and into the appropriate function within the UI.  This was
causing tests to fail since PhantomJS apparently does not have the
Notification class