Add the UI component for the copy-paste app
This adds the actual front-end user-experience components for the
copy-paste workflow.  The structure for the poylmer element is very
similar to the rest of the app: there is an element in the root view
that is hidden when the view mode is not in copy+paste mode.

The connection between the UI and the core is generally handled in the
ui script.  This takes care of displaying the messages as well as the
actions for when the connection is terminated.

For message display, there were a couple different ways I attempted to
structure this, however, they did not work in the end.  The first way I
tried was to delay showing messages until they were completely received,
however, this had some abstraction concerns when checking for
termination conditions.  The second way was to handle batching of
messages on the remote side and send them to the UI all at once,
however, this ran into a problem with some of the messages not making it
through (due to the order of some of the callbacks).