Do not show a notification if a pending shared connection is aborted
If we (in the case, the UI) receives a message to indicate that we
stopped proxying and we did not know we were proxying, do not show any
notification about that.

This also adds tests both for the fact that the notification will be
generated in a normal case and that the notification will not be
generated in the error case.

Fixes #908