Add some accessibility features to the settings pane
The links in the settings pane can now all be accessed through pushing
the tab keys.  Pushing either enter or spacebar when over those links
will now fire the on-tap handler as well.

This also adds a small highlight to the update and reset buttons when
they are focused since paper-button does not do so by default.

Fixes #929