Explicitly checks the view for displaying roster contents
The previous behaviour we were using for the roster was that it should
be displayed anytime that we were not on the splash page, given  that we
had a signed-on network.  This changes that logic so that we will
display it if we are on either the roster page or the settings page,
removing future issues of it appearing on unintended views and fixing
the current issue of it appearing in the copypaste view.