Prevent the user from being in both copypaste and social mode
If the user is currently logged into any networks, we will refuse to
process any links pasted into the browser that should have connection
parameters in them.  Eventually, this should result in an error being
shown to the user, that will be taken care of once we get a separate
page that we load for this situation.

When a copy+paste session is actually active, we will also prevent the
user from being able to navigate back to the splash/roster views.