Remove the 100ms callback from uproxy-tip
We previously had a 100ms callback in uproxy-tip which handled elements
not being fully formed when we first got their position.  Upon further
investigation, this appears to only affect images without the height
defined, so the height of our sharing-enabled image has now been defined
and we no longer have that callback.