Add more logic for the display of "tip"s
This breaks the javascript-based styling of the "tip" window into two
separate parts.  The first part deals with deciding the actual width of
the "tip".  It will be set to half the width of the parent with this
being bounded between 200 and 400 pixels.

The second part determines what side of the "tip" the arrow will be on.
We look at which third of the window it is in (based on the parent's
width) and will put the arrow in the left, middle, or right side based
upon that.