Create a new html page for Chome indicating there is copypaste activity in the main app
Our previous solution for dealing with our inability to completely block
requests for Chorme was to show the actual app window inside the tap.
This caused an issue with any user actions appearing in both the main
app window and the tab.  This switches to showing an extremely simple
(the intent is to update it later with a better UI, just getting this
out for now to show functionality) page that informs the user that the
URL has been received.  The actual app window is opened where we can
show actual information regarding what happened with the link.

The alternate solution would have been to show actual useful information
in the page, however, that created a problem with an inconsistent user
experience between Chrome and Firefox.  In Firefox, I was unable to
redirect the user to an internal resource:// page after intercepting the
page load, so this seems to be the best solution for keeping
functionality consistent between the two apps.