Log the version when starting the app
We now log the version when starting the app, allowing that to be
included in debug data.

This adds a grunt rule to replace the token VERSION in core.ts with a
json-ified object that has the versions we care about in it.

Alternate ideas were having a function (in util.js) that would return
the version and using that for what to print to the log.

Fixes #1036