Improve detection of hidden elements for uproxy-tip
This improves the detection of hidden elements by using
window.getComputedStyle to check whether the current or target element
are currently hidden.  This works in many more cases than the previous
hacky checking of positions.

This also will treat the target element as occupying the full possible
width if there is no parent element (e.g. the previous element is a
top-level component in a web component.