Make sure uproxy tab stays in front with URL pasting
This fixes an issues where the uProxy popup would, in Chrome, not
actually be at the front by the end of the page redirect.  This seems to
be due to the main window regaining focus when the tab loads.

To get around this, we register a callback from that tab to send a
message to the extension telling it to bring the uProxy window to the
front.  This message is handled in the background script for the
extension.

This also moves the responsibility of launching the window out of
generic_ui and into the browser-specific code.

Fixes #1064