Fix positioning of uproxy-bubble (again)
This is (another) attempt to fix the positioning of the uproxy-bubble,
as well as to add some major performance improvements to the element
(i.e. getting rid of most of the polling behaviour).

The first big change is that instead of defining the hidden attribute on
uproxy-bubbles, the recommended use case is now to define the active
attribute.  This essentially functions as !hidden with a few minor
tweaks.  First of all, if the bubble is not active, we do not need to do
any polling and can skip that.  Secondly, we will not actually switch
the bubble to be visible until its target is also visible (we will poll
on that if the bubble is active) meaning when it appears, it will be
correctly positioned.

This makes a relatively simple tweak to the positioning logic where we
remember what width we set the bubble to hold in calculations instead of
looking that up based on offsetWidth, that lets us do the computations
while the bubble is still hidden.