Explicitly set all margins for uproxy-bubble)
The margins on uproxy-bubble all need to be correctly set for the
display to look correct.  Some elements (e.g. core-toolbar) change the
margins for child elements, this will make sure we don't get affected by
that.

Explicitly, and in very boring and non-abstract detail, this fixes the
issues with positioning around the share and settings icons in the
toolbar.