Do a better job handling errors in copy+paste flow
This takes care of making sure that users have a less terrible
experience if anything goes wrong in the copy+paste flow.  It adds error
messages for some conditions we previously did not handle and improves
the routing/behaviour for others.

General page structure:

The page is now broken up into 3 divs: getting, sharing, and (for lack
of a better name) nothing.  Getting and sharing are displayed if the
user is currently in the middle of those processes, nothing is displayed
if the user is on the copy+paste view but does not have an active
connection.  If there is any sharing connection active, it will override
the getting connection and be displayed in its place.  This is a small
departure from the previous structure (each individual state was broken
up and had the conditions applied), it does not have an actual impact on
the display but adds some conceptual structure.

For the actual flow, we will immediately load a link to get access when
the user clicks on the link from the splash page.  If the user pastes a
link requesting access into their browser from that page, or before
visiting that page, we will take them to the page with a link for
sharing access.  This page will have a button to cancel the current
session (clicking on it if you were previously on the page for getting
access will take you back to that same page).  Pasting any link while on
the page to share access is an error.

Once a connection is successfully established, the user is given a page
with a button to end the connection.  Pasting any link while in this
state (with a connection fully established) is considered an error.

Ending a current connection, or having a connection fail unexpectedly
(e.g. NAT issue) will cause the user to be taken to the "nothing" page
which will have an option to request a new connection.  An error will be
shown if it failed unexpectedly.

Errors:

Identifying errors is done in two places: URL parsing and rejection for
the getting-access promise.  If we receive a URL when we are a incorrect
state (though specified above in terms of views, it is actually based on
the getting and sharing state for the underlying remote-connection), an
error informing the user that they need to stop their current connection
is displayed attached to whatever button the user would click to do
that.  If there was a problem with parsing the URL itself, that error is
displayed under the navbar at the top.  If the promise for starting a
getting connection failed, the error will be displayed on the "nothing"
page.  While this may happen in cases where the connection did not
actually fail externally (i.e. we clicked the button to restart the
connection), those cases should not take us to the landing page

Errors are all displayed through uproxy-bubble.