Use core-signal to start copy+paste, require user to start proxying
This has two major changes (really should be two separate commits, but
for once I'm being lazy about this).

First of all, we now use a core-signal to actually trigger switching to
copy+paste mode.  This means we can remove the really ugly functions at
the top of copypaste.ts since we pick up the core signal within the
polymer element.

Second, we no longer immediately start the copy+paste connection but
instead wait for the user to click a button confirming the start.

The pending endpoint (we have to actually establish the proxying
connection before we can tell the user about it) is stored in the
top-level UI object (we need to know about it even if the user closes
the window)

Fixes #1082