Give polymer access to globals in ready functions
This moves the assignment of object properties to global variables to
the ready function of polymer objects instead of being done at the
global level.  This makes sure that we will not have type errors if
the element is loaded without the full UI infrastructure (as long as it
is not actually used).

This also changes the assignment of variables in roster from being in
ready (where we rely on the fact that the javascript references will
hold) to a changed function.