Use more of a stub script for disconnected.html
The disconnected.html now uses a stub script to communicate with the UI
instead of loading the full UI backing.

In Chrome, this is done by sending an extension message which is handled
the same place the copy+paste message is handled.

For Firefox, we inject a content script into the page which listens for
any messages which are either commands or updates.  It then proxies this
up to the glue file which forwards it along to either core or the UI.