Add a handler for clicking on notifications
This adds an onclick handler for all desktop notifications.  It also
adds the optional ability to pass information about what mode clicking
on the notification should switch to as well as which user should be
expanded from clicking on it.  The tag is JSON-serialized, passed into
the tag field for chrome or data for Firefox, and is then passed back to
the UI upon click.