Explicitly update global settings when changing mode
Previously, we had a polymer watcher on model.globalSettings.mode which
would trigger `modeChange` which would update the globalSettings state
in core whenever mode was changed.  This caused `updateGlobalSettings`
to be called multiple times in some cases (except for the two instances
in `root.ts` that relied on the watcher, all other cases still
explicitly updated settings themselves) and resulted in it being called
in error at other points (initialization caused it to be called, even if
globalSettings itself was not fully populated).

We now manually call `updateGlobalSettings` when the user clicks the get
or share tabs.

Fixes #1190