Remove the type errors we see every time we build
The type errors we are seeing when we build actually turn out to be
because freedom should not be called with the new operator, this removes
that.

Initially, we never touch the return value from calling it in
`chrome/app/background.ts`, so don't bother saving it in a variable.

This also fixes the description of freedom in our mock.