Merge instances received from an update
The previous way we were handling receiving instances from an update was
to replace the full array we have with the one we received.  This causes
Polymer to re-create any modified element, which is probably
inefficient.  This also definitely caused a problem when firing an event
with an update occurring in between setting it up and actually calling
it.