Remove some PII from the logs
This removes the most dangerous (and easily-identified) PII from the
logs: emails and pictures.

For emails, we will replace any email address found with EMAIL_{ID}
where ID a numeric ID that will be consistent throughout that single
log.  The original proposal was to use a salted hash with a salt known
only to the reporter, however, that would end up being more difficult to
implement, less performant, and it would offer us a stronger chance of
determining what an email address was based on the logs.

For any base64-encoded image data we detect in the logs, we will replace
it with "IMAGE_DATA".  I cannot imagine any cases where the actual
content of an image would be useful in debugging.

Fixes #1164