core: Wait to clean up old connection before starting a new one
The previous behaviour we had when attempting to start a new connection
was that we would try to terminate any existing connection and would
immediately start trying to connect.  This sometimes caused a race
condition when the state was not completely cleaned up.

To get around that issue, we now wait for the old connections to finish
cleaning up before we try to start a new one.

Fixes #1234