Switch styling for user image in settings to cover
This updates the styling for the user image in settings to be cover,
that should look a bit better for cases where the picture is extra-wide.
Still waiting on getting back user-decided pictures to make this a solid
fix.

This will center the picture which may be undesirable in some cases, any
other positioning also has that possibility though.