Clear old endpoints before setting a new one
Currently, if a user gets disconnected and then tries to reconnect, we
will not clear any old proxying settings on the browser before trying to
set the new ones.  On Chrome, this results in new settings not getting
set and us getting stuck in a state where we are not proxying.

Fixes #1238