Remove the dependency between chrome_core_connector and ui
The chrome core connector previously required access to the ui object in
order to change state when disconnect or connect events occurred.  This
modifies the functionality of the core connector so that it will emit
an event when those events occur.

For firefox, we add a connect function that does nothing except for
emitting the event.