Stop specifying multiple urls per stun server
Due to an upstream bug (willscott/webrtc-adapter#5), we are unable to
specify multiple urls for a single stun server.  To get around this, we
now specify each URL individually.

At the same time, this removes the many different locations where we
have constants for the stun servers; that is now only done in
globals.ts.

Fixes #1332
Fixes #1326