Create and use uproxy-action-dialog
This creates uproxy-action-dialog, a minor extension of
paper-action-dialog.  paper-action-dialog has the possibility of setting
the left and top positions of an element to be <0 which has the
obnoxious side-effect of ruining later calculations of what the width of
an element should be (and which then has the possibility of ruining
later calculations of what the position of the elemnt should be).
uproxy-action-dialog checks that the position set is always on-screen
and corrects that if not.