Try to fix the case where dialogs open too far to the right
The previous fix for dialogs handled the case where we would attempt to
position the dialog to be off the screen which would cause later
calculations for its width to be completely wrong (keeping it off the
screen).  There was another problem we did not notice at the time: that
the second calculation for its position in Firefox would ocasionally
occur before we had any information on what its width would be, set it
to be placed in the middle of the screen, which would end up forcing it
to be offscreen.

Fixes #1361 (again)