Move the functionality to start a proxying session to UserInterface
Starting a connection through a given user is an action that is being
taken place on the level of the entire UI, not specific to a single
instance.  This moves the functionality to actually do that out of the
instance code to the top-level.  By doing this, we are able to get
better checking of whether there was a problem connecting or whether a
user aborted the connection.

To keep track of the current status of our connection, we now use a
variable in the UI that is set to whatever instance we are trying to get
access through.