Sort contacts based on whether getting/giving access to them
On the buddy list, we want to sort contacts based on whether or not we
are getting or giving access to them, with those we are at the top of
the list.

In order to accomplish this, we make a few changes.  First of all, for
the actual sorting, this will look for changes on any of the contacts we
have and re-do the sort.  The actual sort is done by concatenating
onilne and offline contacts and then sorting on whether we are getting
and sharing with the contact followed by the online status.

In order to check for getting/sharing, we needed to have a way to know
what tab was being used.  Historically, we have relied on which tab the
user was currently on to determine what to do, however, this would cause
us to re-sort the contacts every time the user switches tabs; completely
unnecessary considering they live separate from each other.  To get
around this, we now pass the mode that the roster is on down so that we
are able to determine the actual sort order that should be used for a
specific roster-group.

Because of the way polymer watchers work, there is unfortunately no way
to watch for properties of objects in an array changing.  In order to
get around this, we have the uproxy-contact Polymer element fire a
contact-changed event whenever any of the fields we care about change.

Fixes #1308