No longer split contacts by online/offline in the model
Well, we have finally got to the point where we are hacking around our
hack in the model where we split contacts up by whether they are online
or offline, and therefore it is time to remove the hack :).

This allows us to simplify our code all over so, yeah, it's great :).