Open toast message with an event
The previous way we were handling toast messages was to have a flag
(actually the output of a function) that would change depending on
whether the message should be open or not.  This disable the normal ways
of clearing a toast message (e.g. clicking somewhere else or swiping it
offscreen).

We now watch for changes of ui.toastMessage and will fire the open event
when that changes.

Fixes #1333