Factor whether statuses are visible into height calculation
If the statuses at the bottom of the screen are not visible at some
point, we should factor that into the calculation of how tall they are.
In order to do that, we now accept a boolean value into topOfStatuses
that denotes whether or not they can be seen.  If this value is false,
we just use the padding value for the height of the statuses.

In order to check whether they are visible, we also need to figure out
whether or not the feedback window is open (they cannot be seen if it is
open).  To do this, we have it export the attribute opened which is
calculated based off of whether its internal panel is open.

Fixes #1334