Treat welcome bubble as part of header
This starts treating the welcome bubble as part of the header for the
core-header-panel.

For the position of bubbles, we try to do the computation based on the
height of the previous element and its position (prev.height + prev.top
should give us the top amount that we want to use).  This works usually,
but we were running into a problem where the calculation would be done
based on the core-toolbar but the actual positioning would be done based
on a separate div that was placed below core-toolbar due to the selector
in core-header-panel.

In order to fix this, we have added the class core-header to the bubble
which causes it to be filled into the same div in core-header-panel as
the toolbar itself.

Fixes #1523