Add some global css rules
This works towards the goal of getting rid of some of our absolutely
dreadful CSS duplication.  Styles that we find ourselves repeatedly
applying everywhere we use some elements should be contained in the same
place, and we should also be trying to do things with as few rules as
possible (e.g. let's pick a default font size and change it, at the
highest level possible, only where it needs to be changed).

This obviously does not address most of our problems, however, it does
create a single file at the global scope that handles colouring
paper-buttons teal (or grey) as well as well as changing the colouring
for the progress bars we use.

Works towards the conveniently opened #1542