Add a new field, canProxy, to browser APIs
This adds a new field, canProxy, to the browser APIs.  This will be set
to false in cases where the browser is not able to set proxy settings
(in Chrome, this would be when another extension has proxy settings set
and has higher priority).

This also adds an event that will fire in cases where we are proxying
and discover that we has been overriden by another extension.