Show an error when the browser is unable to proxy
If the browser API indicates to the user that we are unable to proxy, we
will now show an error.  If the user is already proxying when this
happens, we will cut off the connection and show them a much more
prominent error.

This also (again) simplifies the logic of calculating the current height
of the statuses.

This also fixes the firing of core signals.

Fixes #1533