Add the ability to the UI to manage multiple accounts
This sets up the ability within the UI to handle multiple accounts.
Primarily, this changes the concept of having a single online network to
storing an array of online networks.  As we get information about
networks, we will add them to this array.  The handling of all update
commands will now lookup the appropriate network to execute the command
under.

There is some minor code cleanup within this as well that switches
setting icons so that it will now be done in a single function.  This
function will check for all conditions that require their own icon and
will set the appropriate icon from that.

In order to make this easier to work with, the model is now a class that
exposes basic functionality such as looking up a network and a user
within a network.

This also updates the tests for UI scripts and attempts to add more unit
functionality to them as well as to remove code duplication.

At this point, no work has been done on the UI, so everything will look
horribly ugly.