Change core to support multiple networks better
This makes a few updates within generic_core to make support for
multiple social networks a bit easier to handle.

First of all, when sending the initial state to the UI, we now return an
array of all currently-logged-in social networks instead of just
returning a single one.  This will make it much easier to handle the
initial connections on the client side.

At the same time as making that change, this also changes the flow of
getting the initial state.  Instead of sending a command and then
receiving an update with the initial state, the initial state is now
sent as the result of the command asking for it.

On the UI side, we have taken out sending the request for initial state
from the various browser connectors and now handle that generic piece of
code within generic_ui/ui.

This also updates the way the global objects in core/social are
stored.  The networks object is now completely managed from within
core/social instead of from core/core.  The pendingNetworks object has
been removed completely and is now a private property of the core.

In order to try to make changing the command work well, the core will
still respond correctly to the old command for this version at least.
When we launch the next version, we can remove all messages noted as
deprecated.