Add icons for the social networks
If the user is logged into multiple social networks, we will now display
an icon indicating which network a user is for next to the avatars.

In order to implement this, we now have a uproxy-avatar element that
handles showing an image (in a way that will "work" with non-square
images) and positioning the svg icon for the network in the bottom-right
corner.  This is used in all positions where we previously used normal
images/divs to show an avatar (i.e. buddylist and settings).

In the case where the picture does not exist, this will show only the
icon for the network.