Make the setup.sh script more robust
This changes the setup.sh script so that we will pass a -f command to rm
in clean (this avoids a failure error code if it is run twice) and
forces us to use the local versions of both bower and grunt giving us
better control over those packages if neded.

This also removes automatically running bower after doing npm install
since the user should, in theory, never run that directly.