Reduce complexity of copy+paste messages
This aims to reduce the complexity of the copy+paste messages as well as
to reduce the computation necessary to process them.

First of all, this creates a new type for the COPYPASTE_MESSAGE message
data, CopyPasteMessage.  It was previously being treated as a
PeerMessage (type was actually {type :PeerMessageType, data :string}),
and with the other updates, the type is now
{type :PeerMessageType, data :PeerMessage[]}.  This removes the need for
us to do manual type casting.

Second, as alluded to above, we now send the peer messages to the UI as
an array of peer messages instead of doing the encoding within the core.
This, and storing them in this fashion within the UI, allows us the
convenience of only needing to pipe the messages through a filter in
order to get the correct output and reduces a bit of complexity there.

Finally, since we are sending the messages down as the array, we no
longer need to handle decoding them back into an object before we add
new messages.