Rename fulfillLaunched to handlePopupLaunch
The name fulfillLaunched was exposing too much information about the
implimentation a single platform (Chrome) had for what to do when a
popup was launched.  This renames that to handlePopupLaunch and defines
that it should be called once the popup is done registering callback.

The time where it is called is moved from whenever the popup receives
a full state update to calling it as soon as all callbacks have been
registered.