Stop exporting a model variable from ui.ts
Currently we export a model variable from ui.ts, this variable is
actually only updated from within the panel context even though it is
avialable globally.  Instead, we now simply have a public property,
model, on our UI object that we can access.

This removes an apparently-functional line from the chrome_tab_auth
script that checked whether we were connected.  Unfortunately, that was
already broken due to the fact that state information never got updated
on the model variable it was referring to, only on the model in the
panel.  There should be no change in functionality there and a fix will
be coming soon (it requires Freedom API changes).