Move categorizeUser to be an unexported function
Previously, we have categorizeUser_ as a private method on the ui object
that would handle moving things in different arrays depending on what
category we wanted a user to be in.  With this change, it is now a
global (unexported) method in the file.  This makes sense because it has
absolutely no tie to class state.