Handle removing networks correctly
Instead of just calling pop on the network array, actually remove
networks when we receive initial state.

The ideal solution here would be to go through each network and find the
difference between them so we only remove the users that are actually
gon.  For the sake of easy implementation, I'm passing on that without
adding a TODO (the advantage that would give us does not seem to be
worth the change).

Fixes #1711