Add better handling of app/extension updates in Chrome
With our current split app/extension design in Chrome, we unfortunately
have to deal with the fact that the two may be of different versions.
In the past, this has been handled by the extension immediately updating
when there was a new version and the app updating whenever the browser
was closed.  This caused us to lose a lot of confidence in them working
together.  In order to fix this, we now handle update messages
ourselves.

When the app has an update available, it will be forwarded to the core
which will store the available version and then forward the information
to the UI.  At this point, the UI will display a notice that an update
is available.  Nothing will happen without user action.

When the extension can update, it will check what the version of the app
is.  If the app is older than the extension, nothing will happen (we
will just update the extension whenever the app updates).  If the app is
newer and we are not currently proxying, we will restart the extension
forcing an update.  This should be transparent to the user.

How could we still have problems?  If something in the Chrome store
messes up and the extension is deployed significantly before the time of
the new app, restarting the browser would still cause our versions to be
out of sync.

Notes on the code you are seeing now (will be updated later)
============================================================

There has not-yet been a thorough test of the functionality changes here
due to the overhead required to do that and the apparent lack of a good
way to test this without making new extensions.

TODO - definitely typed needs to be committed (it does work though once
the pull request gets approved)