Update extension icon after getting initial state
When we get a full state message, update the extension icon after
processing the state.  This ensures that if we previously had bad
information (e.g. thought we were offline) we will have correct
information after finishing this.  This does not fully fix issues with
the icon (we will still show it incorrectly after restarting the
extension before the panel is opened), but it is a step in the right
direction.

Fixes #1717