Add a new StoredValue class to represent a value in storage
The StoredValue class acts as an easy wrapper around a value in storage.
It provides get and set functions that keep a cached value so storage is
not actually accessed every time, and this will guarantee that (assuming
you only used one StoredValue per key) we do not end up with consistency
problems that are too horrible (no locking on read, so you still could
end up with a race condition).