Add a new UI function for updating the view
This adds a new function to the UI, updateView_, that (assuming the UI
is connected to the core) checks all the relevant state we have and
determines what the view should actually be set to based on that.  This
gets called in all the locations we previously set the view.

This change was made because we were often setting the view to either be
splash or roster to signify the state when these may not have been
precisely the most relevant views.

We definitely want to remove the view variable and move just checking
the state components into the Polymer code, but that goes beyond this
quick fix.