Impliment actual reconnection
This takes care of the two remaining steps for reconnection: logging
into networks when the app boots up and storing what networks we should
lag in to.  In order to handle keeping track of the networks, we store a
new key in storage, connectedNetworks, and use this to access the list
of networks.  On initial startup, we go through every value in that
array and immediately try to connect to those networks.

This fails if the Chrome extension is not installed at all, however, it
does seem to work in all other cases.  A new bug has been entered to
buffer outgoing updates when the extension is not installed.

Fixes #1758