Fix issues with core-icon styling
The Chrome handling of dependant styles got changed to give priority to
the parent scope
(https://code.google.com/p/chromium/issues/detail?id=487125) which
causes the styling for core-icon to no longer allow the changing of
height/width using the recommended method.  In order to get around this,
we now spceficy all height/width properties we want to set on a
core-icon as !important (along with some other properties).

Fixes some parts of #1858