Switch to using uproxy-button
One of the downsides of the Chrome rule changes is /deep/ rules will
pretty much universally tend to override other styles on elements.  In
order to get around this, we switch from trying to style all
paper-buttons at a global scope and now have a uproxy-button (that has
styling) and override the styling for that the way we had before where
necessary.

Fixes #1877