copypaste: store the active endpoint and a flag for whether it is in use
Previously, on the UI side, we stored only a pending copy+paste endpoint
and cleared it when it was actually being used.  This had a very weird
mapping to the state that we had in the core (i.e. no mapping).  Instead
of doing that, we now store the endpoint that is currently set up which
has a direct mapping to the core.